requirements.txt
streamlit==1.37.1
pandas==2.3.2
requests==2.32.5
pytz==2025.2
python-dateutil==2.9.0.post0
